# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: zTech CRM API
  description: API del CRM de salones de eventos corporativos — UNLaM GADS1.
  version: v1
servers:
- url: http://localhost:53103
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/v1/contacts/{id}:
    get:
      tags:
      - Contacts
      summary: Detalle de un contacto
      description: Para SELLER exige asignación directa o una oportunidad propia relacionada.
      operationId: getById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactDetailResponse"
        "404":
          description: No existe un contacto con ese id
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactDetailResponse"
    put:
      tags:
      - Contacts
      summary: Edita un contacto
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
        required: true
      responses:
        "200":
          description: Contacto actualizado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: "No existe un contacto con ese id, o la empresa relacionada\
            \ no existe"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "409":
          description: Ya existe un contacto con ese documento
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /api/v1/companies/{id}:
    get:
      tags:
      - Companies
      summary: "Detalle de una empresa, con sus contactos (BE-CUS-03)"
      description: Para SELLER exige asignación directa o una oportunidad propia;
        los contactos anidados también se filtran por ese alcance.
      operationId: getById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyDetailResponse"
        "404":
          description: No existe una empresa con ese id
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyDetailResponse"
    put:
      tags:
      - Companies
      summary: Edita una empresa
      operationId: update_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyRequest"
        required: true
      responses:
        "200":
          description: Empresa actualizada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "404":
          description: No existe una empresa con ese id
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "409":
          description: Ya existe una empresa con ese CUIT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
...
  /api/v1/contacts:
    get:
      tags:
      - Contacts
      summary: "Lista contactos visibles, paginado"
      description: ADMIN y SALES_MANAGER ven el tenant completo. SELLER ve asignaciones
        directas y contactos relacionados con oportunidades propias.
      operationId: list_2
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
      - name: originId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: salesRepId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: companyId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "lastName,asc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseContactResponse"
    post:
      tags:
      - Contacts
      summary: Crea un contacto
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
        required: true
      responses:
        "201":
          description: Contacto creado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: La empresa relacionada no existe
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "409":
          description: Ya existe un contacto con ese documento
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /api/v1/companies:
    get:
      tags:
      - Companies
      summary: "Lista empresas visibles, paginado"
      description: ADMIN y SALES_MANAGER ven el tenant completo. SELLER ve asignaciones
        directas y empresas relacionadas con oportunidades propias.
      operationId: list_3
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
      - name: originId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: salesRepId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "businessName,asc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseCompanyResponse"
    post:
      tags:
      - Companies
      summary: Crea una empresa
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyRequest"
        required: true
      responses:
        "201":
          description: Empresa creada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "409":
          description: Ya existe una empresa con ese CUIT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
...
components:
  schemas:
...
    ContactRequest:
      type: object
      properties:
        companyId:
          type: integer
          format: int64
          description: Id de la empresa relacionada. Vacío si es cliente individual
            (Módulo 1 de la consigna).
        firstName:
          type: string
          example: María
          minLength: 1
        lastName:
          type: string
          example: Gómez
          minLength: 1
        document:
          type: string
          description: Opcional
          example: "30123456"
        position:
          type: string
          description: "Cargo, si corresponde"
          example: Responsable de compras
        email:
          type: string
          format: email
          example: maria.gomez@empresa.com
        phone:
          type: string
          example: +54 11 4000-0000
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
          description: "Id del usuario responsable comercial, si ya está asignado"
        originId:
          type: integer
          format: int64
          description: Id del origen comercial
        notes:
          type: string
      required:
      - firstName
      - lastName
      - status
    ContactResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
    CompanyRequest:
      type: object
      properties:
        legalName:
          type: string
          example: Eventos del Sur SA
          minLength: 1
        businessName:
          type: string
          example: Eventos del Sur
          minLength: 1
        cuit:
          type: string
          description: "Opcional. Formato: 11 dígitos, con o sin guiones."
          example: 30-71234567-9
        industry:
          type: string
          example: Turismo y eventos
        email:
          type: string
          format: email
          example: contacto@eventosdelsur.com
        phone:
          type: string
          example: +54 11 4000-0000
        address:
          type: string
          example: "Av. Rivadavia 4500, San Justo"
        locality:
          type: string
          example: San Justo
        website:
          type: string
          example: https://eventosdelsur.com
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
          description: "Id del responsable. Si se omite, el backend aplica la herencia\
            \ de DP-02"
        originId:
          type: integer
          format: int64
          description: Id del origen comercial
        notes:
          type: string
      required:
      - businessName
      - legalName
      - status
    CompanyResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        legalName:
          type: string
        businessName:
          type: string
        cuit:
          type: string
        industry:
          type: string
        email:
          type: string
        phone:
          type: string
        locality:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
...
    CompanySummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        legalName:
          type: string
        businessName:
          type: string
        salesRepId:
          type: integer
          format: int64
    ContactSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        companyId:
          type: integer
          format: int64
        salesRepId:
          type: integer
          format: int64
...
    PageResponseContactResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ContactResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    ContactDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        company:
          $ref: "#/components/schemas/CompanySummaryResponse"
        firstName:
          type: string
        lastName:
          type: string
        document:
          type: string
        position:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
        originId:
          type: integer
          format: int64
        notes:
          type: string
    PageResponseCompanyResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CompanyResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    CompanyDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        legalName:
          type: string
        businessName:
          type: string
        cuit:
          type: string
        industry:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        locality:
          type: string
        website:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
        originId:
          type: integer
          format: int64
        notes:
          type: string
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/ContactSummaryResponse"
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT